requirements.txt
Flask
gunicorn
beautifulsoup4
requests
imageio-ffmpeg
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
